light/dark toggle refactor, added JS module files
