+++ data.js
@@ -1,0 +1,15 @@
+let productA = {
+    id: 1,
+    emoji: "⭐",
+    revenue: 200,
+    commission: 50
+};
+
+let productB = {
+    id: 2,
+    emoji: "🔥",
+    revenue: 300,
+    commission: 75
+};
+
+export { productA, productB }
