+++ index.html
@@ -144,7 +144,7 @@
         </main>
     </div>
 
-    <script src="index.js"></script>
+    <script src="index.js" type="module"></script>
 
 </body>
 </html>
